Add system parameter requests & model state spec
Added `SystemABC.requested_parameters()/model_state()` to allow systems
to provide information about their underlying state specification,
including shape, as well as request parameters that comply with said
requirements. Also integrated this work into the existing wrapper system
to demonstrate the functionality.

- Added `requested_parameters()/model_state()` methods to `SystemABC`. 
- Extended `wrap()`/`_AdapterSystem` to accept explicit parameter
  request and model state callbacks.
- Updated `_checked_partial()` so already materialized `ParameterValue`
  instances are preserved when binding static parameters.
- Simplified the documentation quick start wrapper config to the minimal
  valid example.

+++ tests/system/system_wrapper_assets/dummy_system.py
@@ -1,11 +1,13 @@
 """A dummy stepper function for testing `WrapperSystem`."""
 
-import numpy as np
-
 from flepimop2.typing import Float64NDArray
 
 
-def stepper(time: float, state: Float64NDArray, offset: np.float64) -> Float64NDArray:
+def stepper(
+    time: float,
+    state: Float64NDArray,
+    offset: float,
+) -> Float64NDArray:
     """
     A dummy stepper function for testing purposes: (state + offset) * time.
